+++ enterprise/j2eeserver/src/org/netbeans/modules/j2ee/deployment/impl/ui/ProgressUI.java
@@ -91,7 +91,7 @@ public class ProgressUI implements ProgressListener {
             progressComponent = ProgressHandleFactory.createProgressComponent(handle);
         }
         if (delay != null) {
-            handle.setInitialDelay(delay.intValue());
+            handle.setInitialDelay(delay);
         }        
         handle.start();
     }
